consistent organization
